Fixes Issue #287
Had to change a bunch around because there where a few errors where methods got changed in a very sloppy way. i have fixed them and added one that checks for this issue and prevents it. in doing so i needed to fix how the rooms combboxes were being refered to and i have two strings loaded and ready to go but i cant get the combo boxes to select properly.